requirements.txt
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
typing_extensions==4.12.2